New Slimmer Endorsed tarball
These jars represent the jars needed to be present in the thredds'
WEB-INF lib such that it works with the ORP properly.  Specifically
to have the security filters from this project that are installed on
the TDS to work.  So this file went on a big diet.

Email form Luca:
"The thredds directory WEB-INF/lib is missing the following jars (which
were previously in the endorsed directory):
                - xmlsec-1.4.2.jar
                - commons-collections-3.2.1.jar
                - velocity-1.5.jar
                - xalan-2.7.1.jar
                - xercesImpl-2.9.1.jar
                - xml-apis-2.9.1.jar
                - serializer-2.9.1.jar"

The only difference is that the xml-apis is a newer version, but the
same as the version used in the project to date: v1.3.04