The idea is not to have to set a property file for the ORP to use but have it fetch the data it needs directly by reading the server.xml file.  This lowers maintenance, because if the user changes keystore or key alieas or password then the orp will automatically stay in step without intervention from the end user.
Estani wrote code to pull out the data from the server.xml file into a
property object.  But then the code that is to use this is a spring
injected object.  I think that these two approaches are incongruent.
I don't mess with spring so then Luca is going to have to look at it
with Estani to straighten this thing out.  Of course we want to cut a
release now.

"Murphy's Law is still in effect, ain't that right wood." -3xD