Will not overwrite esgf_idp_static.xml file if present
Luca noticed:
a) the file esgf_idp_static.xml is overwritten every time the node
starts. This is bad if the file has been manually edited by the local
admin to allow other IdPs - could you confirm, and if so stop this
from happening ?

Indeed it was being written if changed.  So, to satisfy the new
semantics we are not going to overwrite the file if one is present.