Using distribution files that had been moved over from security
Luca's commit 97913bc (two commits prior) moved over files that were
once in security, over into the orp.  This supports the data node only
configuration since it was the security module that was the coupling
between the IDP and the DATA configuration types.

I moved over the configuration logic in the install script to use
these files accordingly.